Fix the Assertions of the Identifier(s) value of Type CoNE
-> If an identifier of type CoNE is set in an item the identifier value
of the saved item is a CoNE link (not the single value/id anymore)
=> In the data driven tests using CoNE as Identifier(s) Type the
assertion of the Identifiers value must use a CoNE-URL link for
comparison

- Add TestHelper.compareIdentifiers() method to compare a given
identifier(s) value with the actual identifier(s) value. If the
identifier type is 'CoNE', then the corresponding CoNE-URL links will be
compared (not the single value/id).
- Adapt the data driven tests, which set the identifier(s) field in
their items